Added Company controller, Dto, Model and Services so as to create a new Company

+++ Src/Contact/Service/ContactService.cs
@@ -1,10 +1,9 @@
 using ContactApp.Data;
-using ContactApp.Dtos;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace ContactApp.Src.Contact.Service;
 
+using ContactApp.Src.Contact.Dtos;
 using ContactApp.Src.Contact.Model;
 
 public class ContactService
@@ -18,10 +17,23 @@ public class ContactService
     {
         return _context.Contact.AsNoTracking().ToList().Select(contact => contact.AsDto());
     }
-    public ContactDto Get(int id)
+    public ContactDto? Get(int id)
     {
         var value = _context.Contact.AsNoTracking().SingleOrDefault(p => p.Id == id);
         return value.AsDto();
+    }
+    public async Task AddContactAsync(ContactDto contactDto)
+    {
+        var contact = new Contact
+        {
+            Id = contactDto.Id,
+            Number = contactDto.Number,
+            Provider = contactDto.Provider
+        };
+
+        _context.Contact.Add(contact);
+        await _context.SaveChangesAsync();
+
     }
     public async Task<ContactDto> AddNewContact(CreateContactDto contactDto)
     {
@@ -72,11 +84,16 @@ public class ContactService
         await _context.SaveChangesAsync();
     }
 
-    public void DeleteContact(int id)
+    public bool DeleteContact(int id)
     {
         var contact = _context.Contact.Find(id);
-         _context.Contact.Remove(contact);
+        if (contact is null)
+        {
+            return false;
+        }
+        _context.Contact.Remove(contact);
         _context.SaveChanges();
+        return true;
     }
